requirements.txt
flask
sqlalchemy
flask_sqlalchemy
flask_script
oauth2client
requests
pyjwt
flask_cors
jsonschema
slugify